WIP: Render the whole view instead of messageBody
